fix(UnityIntegration): Unity .meta files for releases
The Unity Package Manager handles dependencies by fetching them as a
read-only copy. Since Unity needs a `.meta` file for each file found
in the package the package will be unusable as the package folder is
read-only. The fix is to ensure the released package contains all
the necessary `.meta` files.