fix(SerializedProperty): help serializing properties, not fields
Previously the `FieldToProperty` feature helped creating properties
for a field to simplify the handling of changes to the field in the
Unity inspector. While it was a useful feature it turns out that
supporting any usage of that field results in the need to perform
the weaving process on any assembly, at all times. This is primarily
an issue when a field gets "changed" into a property in an assembly
A while assembly B contains a usage of that field, which is no
longer accessible. Instead of doing a reverse lookup when weaving
assembly B to find the usage of a "changed" field in assembly A this
fix changes the feature completely: Instead of turning fields into
properties properties will be serialized by Unity instead. This also
allows the serialized member to now use any accessible, including
different accessibility for the property's getter vs. its setter.
This new feature supports any backing field, including
auto-implemented ones for which an automatic renaming of the field
ensures the Unity inspector shows the property by its human-readable
name.

BREAKING CHANGE: Fields can no longer be "changed" into properties.
Instead of using a field and annotating it with
`FieldToPropertyAttribute` use a property and annotate it with
`SerializedPropertyAttribute`. Auto-implemented properties as well
as regular ones are supported and the only requirement is to have a
getter and setter (of any accessibility level).