refactor(SerializedProperty): invert the hidden-in-inspector setting
Boolean members should be written in a way that they are positive,
not negative. This changes the attribute constructor and the field
to adhere to that convention.

BREAKING CHANGE: The argument passed in the `SerializedProperty`
attribute constructor to hide the field in the inspector is now
negated. Uses need to be updated to pass the negation of what they
previously passed. The default value of the parameter has been
updated which means the default constructor call doesn't need to be
changed to upgrade to this change.