feat(UnityIntegration): don't implicitly reference any plugin .dll
Unity allows to set assemblies to be explicitly referenced only.
This is now turned on to reduce load times as well as build sizes.
The runtime assemblies that include the attributes to use still are
implicitly referenced. This allows projects that don't use Assembly
Definition files to use the attributes without issues.

BREAKING CHANGE: Referencing a Malimbe Editor assembly has to be
done explicitly from now on. Use Assembly Definition Files and their
inspector to reference the needed assemblies manually.