feat(BehaviourStateRequirement): allow returning early in Behaviours
In a `UnityEngine.Behaviour` or a type derived from it it is often
useful to return early while the Behaviour or the GameObject the
Behaviour is on are disabled/inactive. This new weaver takes care of
adding the necessary instructions to a method to return early based
on those two states.