docs(structure): adhere to unity technologies branding guidelines
The use of the wordmark `Unity` has been updated in the README to
clearly adhere to the guidelines laid out by Unity Technologies at
https://unity3d.com/legal/branding_trademarks and a disclaimer
has been added to clearly outline that this repository is not in
affiliation with Unity Technologies.

Any reference to VRTK has been updated to reference the Extend
Reality Ltd company entity as this is the appropriate governing
body of the repository.

The CONTRIBUTING.md has also been updated to remove the unnecessary
72 character line limit.

The LICENSE.md file has also been updated to extend the copyright
to 2019 as changes have been made in this year as well.