feat(PropertySetter): allow using property inside called method
If the annotated method tried to call another method that uses the
property it would end up working on the previous value instead of
whatever this method returns. This change allows to set the value of
the property's backing field at any time, thus it can be set before
calling a method that needs the latest property value by calling the
property's getter.

BREAKING CHANGE: `SetsPropertyAttribute` was renamed to
`CalledBySetterAttribute` and the expected signature of the
annotated method changed. Please see the latest Readme for details.