fix(UnityIntegration): delay weaving process until Unity is ready
On startup of a new Unity project (or when one deleted the project's
`Library` folder) Unity was logging hundreds of errors because of
Malimbe's Editor integration running before the needed subsystems of
Unity are initialized. The fix is to delay the weaving process until
the Editor has updated all inspectors once.