fix(UnityIntegration): tell Unity to reload assemblies when needed
To handle the startup behavior of Unity a delay was added to only
start the weaving process once Unity is initialized properly. That
change introduced an issue: When Malimbe weaves the assemblies Unity
won't know about it and as such, even though the assembly on disk
was changed by Malimbe, Unity still uses the unchanged version it
compiled. The fix is to ask Unity to do a reload of the weaved
assemblies.