fix(UnityIntegration): prevent serialization issues
If a Unity project wasn't yet using Malimbe and runs it for the
first time during the startup of the Unity editor Unity won't pick
up changes of serialized values correctly until the serialized asset
was touched at least once. The fix is to revert the integration
behavior so it runs the weaving process on startup. This brings back
an old issue: The Unity console will be logging issues regarding
"Unity extensions" hundreds of times. To prevent this the Unity API
that logs those errors is no longer called instead. No behavioral
change has been observed and as such this change is not labeled a
breaking change.