refactor(PropertyValidation): remove property validation
Property validation doesn't actually work in the intended sense:
Simply setting a property to its current value is missing the
ability to act on the previous value. As such the need for the
validation feature offered by Malimbe is gone. This change removes
it.

BREAKING CHANGE: Property validation has been removed from Malimbe.
For an alternative solution please look out for a future version of
this project or
[Zinnia](https://github.com/ExtendRealityLtd/Zinnia.Unity).