feat(MemberChange): only call change methods when isActiveAndEnabled
For Behaviours it is important to ignore changes while the Behaviour
is disabled as it should handle the state in its OnEnable
implementation. This new feature ensures the change handler methods
are only called for Behaviours that are active and enabled. The
change can be done by either using the inspector or calling a
property's setter.