feat(UnityIntegration): allow customizing InspectorEditor
The InspectorEditor is used to call change handlers. As this
functionality is crucial even when a custom inspector for a type has
to be implemented this change ensures the editor is customizable
through subclassing.

This change also ensures no undo operations are written to the Undo
stack as this inspector can't call the change handler methods when
undoing or redoing changes.