feat(MemberChange): help preventing infinite loops
When a property is changed the MemberChange feature of Malimbe will
call annotated methods before or after the change. Changing the
property in those annotated change handler methods will in turn call
the setter, thus recursively call the handler methods again. To
prevent this issue in most cases this change ensures that property
setter calls in the change handler methods instead store the value
into the backing field directly. While this change doesn't help when
the change handler methods call other methods that ultimately call
the property setter this is still a good help for the use case of
validating a property.