fix(MemberChange): allow undo operations but warn against using them
Supporting the Undo stack in Unity is important as a lot of logic
depends on it, most importantly the dirty checks for the scene or a
prefab that was modified. To ensure the expected behavior is
followed this change allows any change to be added to the Undo stack
and instead warns users against running undo/redo operation of those
changes because they won't be handled by change handler methods.