fix(MemberChange): refactor inspector logic to be more composable
The previous refactor of the custom unity InspectorEditor did not
split out enough of the logic into their own methods so they were
not completely usable independently.

This refactor splits out the methods even more so they are more
independent and can be easily used on their own or overridden.